requirements.txt
# Core
python-dotenv
pyyaml
loguru

# Data + distributed processing
pandas>=2.0
dask[dataframe]>=2023.3
pyarrow

# Geo / parsing
geohash2

# ML
scikit-learn==1.5.2
lightgbm

# API
fastapi
uvicorn
httpx

# Monitoring
prometheus-client

# Testing
pytest
pytest-cov

# Orchestration
#apache-airflow==2.8.2

# Misc
email-validator
dill